sing/unigned compare warning was enabled for gcc/clang and made usable, close #507

+++ solvers/optical/slab/fourier/solver2d.cpp
@@ -140,7 +140,7 @@ void FourierSolver2D::loadConfiguration(XMLReader& reader, Manager& manager)
 void FourierSolver2D::onInitialize()
 {
     this->setupLayers();
-    if (this->interface == size_t(-1))
+    if (this->interface == -1)
         Solver::writelog(LOG_DETAIL, "Initializing Fourier2D solver ({0} layers in the stack)",
                                      this->stack.size());
     else
